+++ Sources/NSManagedObjectContext+Utils.swift
@@ -196,7 +196,7 @@ extension NSManagedObjectContext {
   ///
   /// Synchronously performs a given block on the context’s queue and returns the final result.
   /// - Throws: It throws an error in cases of failure.
-  public func performAndWait<T>(_ block: (NSManagedObjectContext) throws -> T) rethrows -> T {
+  public func performAndWaitResult<T>(_ block: (NSManagedObjectContext) throws -> T) rethrows -> T {
     return try _performAndWait(function: performAndWait, execute: block, rescue: { throw $0 })
   }